src/requirements.txt
python-dotenv==1.1.1
pydantic==2.11.9
openai==1.109.1
Flask==3.1.2
psycopg2-binary==2.9.10
pytz==2025.2
